+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.3.0] 2025-07-30
+
+### Changed
+
+- ML_callbacks: change `np.Inf` to `np.inf`, requiring Numpy version >= 2.0.
+- Fixed dependency `Numpy>=2.0` for the core Machine Learning modules.
+- utilities: Use module logger instead of print.
+- data_exploration: Use module logger instead of print.
+
+### Fixed
+
+- ML_datasetmaster, correct dtype when creating datasets for regression tasks.
+
 ## [5.2.2] 2025-07-28
 
 ### Added